// HealthCareApiServer/WebApp/client-app/src/app/api/models/prescription-model.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { EntityModel } from '../models/entity-model';
import { PrescriptionIntakeModel } from '../models/prescription-intake-model';
export type PrescriptionModel = EntityModel & {
'name': string;
'dose'?: string | null;
'patientId': number;
'appointmentId'?: number;
'startDate'?: string;
'durationInDays'?: number;
'times'?: Array<string> | null;
'isActive'?: boolean;
'intakes'?: Array<PrescriptionIntakeModel> | null;
};